Updating database credential keys for statistics
tools/lsrd_stats.py is using the dbnfo values which are currently only h for host, u for user, p for password and d for database.  This configuration has changed on the system so the stats are generated from the historical database.  Multiple other scripts in tools/ are pointed to this cred file including order_monitoring, which needs to look only at the ordering database.

had to create a separate set of credentials in dbnfo specifically for lsrd_stats.py.